requirements.txt
networkx
torch
torchvision
torch-summary
einops
scikit-image
opencv-python
pyclipper
shapely
requests
oscrypto 
freetype-py
googletrans==4.0.0rc1
aiohttp
tqdm
sklearn
deepl
ImageHash
kornia
python-dotenv
backports.cached-property